+++ src/output.F90
@@ -3,7 +3,7 @@ Module  output
    use stdlib_io_npy, only:load_npy,save_npy
    use Class_helper
    implicit None
-   integer(4), parameter  :: std_out =  6
+   integer(int64), parameter  :: std_out =  6
    interface print_mtx
       module procedure print_mtx_real, print_mtx_cmplx, print_vec_real,&
          print_mtx_real_no_unit, print_mtx_cmplx_no_unit, &
@@ -14,16 +14,16 @@ Module  output
 contains
    subroutine print_mtx_int_no_unit(mtx)
       Implicit None
-      integer(4), dimension(:,:), intent(in)    :: mtx
+      integer(int64), dimension(:,:), intent(in)    :: mtx
 
       call print_mtx_int(std_out,  mtx)
    end subroutine
 
    Subroutine  print_mtx_int(p_unit, mtx)
       Implicit None
-      integer(4), dimension(:,:), intent(in)    :: mtx
-      integer(4), intent(in)                 :: p_unit
-      integer(4)                             :: i,j
+      integer(int64), dimension(:,:), intent(in)    :: mtx
+      integer(int64), intent(in)                 :: p_unit
+      integer(int64)                             :: i,j
 
       do i = 1, size(mtx,1)
          do j =  1, size(mtx,2)
@@ -38,16 +38,16 @@ contains
 
    subroutine print_vec_int_no_unit(vec)
       implicit none
-      integer(4), dimension(:), intent(in)   :: vec
+      integer(int64), dimension(:), intent(in)   :: vec
 
       call print_vec_int(std_out, vec)
    end subroutine print_vec_int_no_unit
 
    subroutine print_vec_int(p_unit, vec)
       implicit none
-      integer(4), dimension(:), intent(in)   :: vec
-      integer(4), intent(in)                 :: p_unit
-      integer(4)                             :: i
+      integer(int64), dimension(:), intent(in)   :: vec
+      integer(int64), intent(in)                 :: p_unit
+      integer(int64)                             :: i
 
       do i =  1,size(vec)
          write(p_unit,"(I6)") vec(i)
@@ -56,16 +56,16 @@ contains
 
    subroutine print_vec_real_no_unit(vec)
       implicit none
-      real(8), dimension(:), intent(in)      :: vec
+      real(dp), dimension(:), intent(in)      :: vec
 
       call print_vec_real(std_out,  vec)
    end subroutine print_vec_real_no_unit
 
    subroutine print_vec_real(p_unit, vec)
       implicit none
-      real(8), dimension(:), intent(in)      :: vec
-      integer(4), intent(in)                 :: p_unit
-      integer(4)                             :: i
+      real(dp), dimension(:), intent(in)      :: vec
+      integer(int64), intent(in)                 :: p_unit
+      integer(int64)                             :: i
 
       do i =  1, size(vec)
          write(p_unit, "(ES10.3)") vec(i)
@@ -74,16 +74,16 @@ contains
 
    subroutine print_vec_cmplx_no_unit(vec)
       implicit none
-      complex(8), dimension(:), intent(in)      :: vec
+      complex(dp), dimension(:), intent(in)      :: vec
 
       call print_vec_cmplx(std_out,  vec)
    end subroutine print_vec_cmplx_no_unit
 
    subroutine print_vec_cmplx(p_unit, vec)
       implicit none
-      complex(8), dimension(:), intent(in)      :: vec
-      integer(4), intent(in)                 :: p_unit
-      integer(4)                             :: i
+      complex(dp), dimension(:), intent(in)      :: vec
+      integer(int64), intent(in)                 :: p_unit
+      integer(int64)                             :: i
 
       do i =  1, size(vec)
          write(p_unit, "(A, ES10.3, A, ES10.3, A)")&
@@ -93,16 +93,16 @@ contains
 
    subroutine print_mtx_real_no_unit(mtx)
       Implicit None
-      real(8), dimension(:,:), intent(in)    :: mtx
+      real(dp), dimension(:,:), intent(in)    :: mtx
 
       call print_mtx_real(std_out,  mtx)
    end subroutine
 
    Subroutine  print_mtx_real(p_unit, mtx)
       Implicit None
-      real(8), dimension(:,:), intent(in)    :: mtx
-      integer(4), intent(in)                 :: p_unit
-      integer(4)                             :: i,j
+      real(dp), dimension(:,:), intent(in)    :: mtx
+      integer(int64), intent(in)                 :: p_unit
+      integer(int64)                             :: i,j
 
       do i = 1, size(mtx,1)
          do j =  1, size(mtx,2)
@@ -117,16 +117,16 @@ contains
 
    subroutine print_mtx_cmplx_no_unit(mtx)
       Implicit None
-      complex(8), dimension(:,:), intent(in)    :: mtx
+      complex(dp), dimension(:,:), intent(in)    :: mtx
 
       call print_mtx_cmplx(std_out,  mtx)
    end subroutine print_mtx_cmplx_no_unit
 
    Subroutine  print_mtx_cmplx(p_unit, mtx)
       Implicit None
-      complex(8), dimension(:,:), intent(in)    :: mtx
-      integer(4), intent(in)                     :: p_unit
-      integer(4)                                 :: i,j
+      complex(dp), dimension(:,:), intent(in)    :: mtx
+      integer(int64), intent(in)                     :: p_unit
+      integer(int64)                                 :: i,j
       character(len=4)                           :: i_str
       character(len=4)                           :: deli = "j), "
 
@@ -152,7 +152,7 @@ contains
    subroutine create_dir(folder)
       implicit none
       character(len=*) :: folder
-      integer(4)       :: succ
+      integer(int64)       :: succ
 
       call execute_command_line("rm " // trim(folder) // "*.npy", exitstat=succ)
       if(succ /= 0) then
